Move media session related code out of MainActivity
